odw/services/data-warehouse/app/seed-data/requirements.txt
Pillow>=10.0.0
boto3>=1.26.0
toml>=0.10.0 